refactor basic form
